requirements.txt
aiogram==2.15
python-dotenv==1.0.1
APScheduler==3.10.4
asyncpg==0.29.0